Split MinimumSpanningTree header to manage complexity
